+++ src/toParts.ts
@@ -1,10 +1,10 @@
 import {
 	DAYS_IN_WEEK,
 	HOURS_IN_DAY,
-	MICROSECONDS_IN_A_MILLISECOND,
-	MILLISECONDS_IN_A_SECOND,
+	MICROSECONDS_IN_MILLISECOND,
+	MILLISECONDS_IN_SECOND,
 	MINUTES_IN_HOUR,
-	NANOSECONDS_IN_A_MICROSECOND,
+	NANOSECONDS_IN_MICROSECOND,
 	SECONDS_IN_MINUTE,
 } from './constant.js';
 import {
@@ -44,12 +44,12 @@ export const toParts = (ms: number): Parts => {
 		hours: Math.floor(hours.from(absoluteMs) % HOURS_IN_DAY),
 		isNegative: ms < 0,
 		nanoseconds: Math.round(
-			nanoseconds.from(absoluteMs) % NANOSECONDS_IN_A_MICROSECOND,
+			nanoseconds.from(absoluteMs) % NANOSECONDS_IN_MICROSECOND,
 		),
 		microseconds: Math.round(
-			microseconds.from(absoluteMs) % MICROSECONDS_IN_A_MILLISECOND,
+			microseconds.from(absoluteMs) % MICROSECONDS_IN_MILLISECOND,
 		),
-		milliseconds: Math.round(absoluteMs % MILLISECONDS_IN_A_SECOND),
+		milliseconds: Math.round(absoluteMs % MILLISECONDS_IN_SECOND),
 		minutes: Math.floor(minutes.from(absoluteMs) % MINUTES_IN_HOUR),
 		seconds: Math.floor(seconds.from(absoluteMs) % SECONDS_IN_MINUTE),
 		weeks: Math.floor(weeks.from(absoluteMs)),