+++ src/types/TimeUnit.ts
@@ -1,1 +1,14 @@
+/**
+ * Represents all possible time units.
+ *
+ * - `ns` - Nanoseconds
+ * - `μs` - Microseconds
+ * - `ms` - Milliseconds
+ * - `s` - Seconds
+ * - `m` - Minutes
+ * - `h` - Hours
+ * - `d` - Days
+ * - `w` - Weeks
+ */
+
 export type TimeUnit = 'ns' | 'μs' | 'ms' | 's' | 'm' | 'h' | 'd' | 'w';