+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.2] - 2025-05-06
 
-### Changes
+### Changed
 
 - Workflow adjustments (thanks [@simmo])