+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Incorrect/missing documentation (thanks [@simmo])
+- Floating point conversion error (thanks [@simmo])
 
 ## [1.2.0] - 2025-06-10